refactor(runtime): Rehome bootstrap config glue
Move NodeBootstrap and NodeConfigManager into network.crypta.runtime.bootstrap.

Update Node, affected tests, and SpotBugs excludes to use the new package without changing startup or configuration behavior.
