+++ src/main/java/network/crypta/node/Node.java
@@ -34,12 +34,12 @@ import network.crypta.l10n.NodeL10n;
 import network.crypta.node.SecurityLevels.NETWORK_THREAT_LEVEL;
 import network.crypta.node.SecurityLevels.PHYSICAL_THREAT_LEVEL;
 import network.crypta.node.subsystem.CryptoAndTransportParams;
-import network.crypta.node.subsystem.NodeBootstrap;
-import network.crypta.node.subsystem.NodeConfigManager;
 import network.crypta.node.subsystem.NodeMessagingSubsystem;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
 import network.crypta.node.subsystem.NodeRoutingSubsystem;
 import network.crypta.node.subsystem.NodeStorageSubsystem;
+import network.crypta.runtime.bootstrap.NodeBootstrap;
+import network.crypta.runtime.bootstrap.NodeConfigManager;
 import network.crypta.runtime.endpoints.NodeClientCoreInit;
 import network.crypta.runtime.services.NodeServicesSubsystem;
 import network.crypta.support.Fields;
