+++ src/main/java/network/crypta/runtime/bootstrap/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * Startup bootstrap and config wiring helpers for early node initialization.
+ *
+ * <p>This package groups the neutral runtime helpers that prepare program directories, random
+ * sources, and node configuration callbacks during startup. The classes remain part of the daemon
+ * runtime and preserve their existing behavior, but they are not part of the routing, storage, or
+ * message kernel.
+ */
+package network.crypta.runtime.bootstrap;
