+++ src/test/java/network/crypta/node/AnnounceSenderTest.java
@@ -6,8 +6,8 @@ import network.crypta.io.comm.Message;
 import network.crypta.io.comm.MessageCore;
 import network.crypta.io.comm.MessageFilter;
 import network.crypta.io.comm.NotConnectedException;
-import network.crypta.node.subsystem.NodeBootstrap;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
+import network.crypta.runtime.bootstrap.NodeBootstrap;
 import network.crypta.support.PriorityAwareExecutor;
 import network.crypta.support.io.NativeThread;
 import org.junit.jupiter.api.BeforeEach;
