+++ src/test/java/network/crypta/node/DarknetPeerNodeTest.java
@@ -8,8 +8,8 @@ import network.crypta.io.comm.Message;
 import network.crypta.io.comm.Peer;
 import network.crypta.node.DarknetPeerNode.FRIEND_TRUST;
 import network.crypta.node.DarknetPeerNode.FRIEND_VISIBILITY;
-import network.crypta.node.subsystem.NodeBootstrap;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
+import network.crypta.runtime.bootstrap.NodeBootstrap;
 import network.crypta.support.Base64;
 import network.crypta.support.SimpleFieldSet;
 import network.crypta.support.math.MersenneTwister;
