+++ src/test/java/network/crypta/node/NodeAnnouncementHandlerTest.java
@@ -5,8 +5,8 @@ import java.util.stream.Stream;
 import network.crypta.io.comm.DMT;
 import network.crypta.io.comm.Message;
 import network.crypta.io.comm.MessageType;
-import network.crypta.node.subsystem.NodeBootstrap;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
+import network.crypta.runtime.bootstrap.NodeBootstrap;
 import network.crypta.support.PriorityAwareExecutor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
