+++ src/test/java/network/crypta/node/NodeOfferMessageHandlerTest.java
@@ -10,9 +10,9 @@ import network.crypta.io.comm.NotConnectedException;
 import network.crypta.keys.Key;
 import network.crypta.keys.NodeSSK;
 import network.crypta.node.NodeStats.RejectReason;
-import network.crypta.node.subsystem.NodeBootstrap;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
 import network.crypta.node.subsystem.NodeRoutingSubsystem;
+import network.crypta.runtime.bootstrap.NodeBootstrap;
 import network.crypta.support.ShortBuffer;
 import network.crypta.support.Ticker;
 import org.junit.jupiter.api.BeforeEach;
