+++ src/test/java/network/crypta/node/simulator/RealNodePitchBlackMitigationTestTest.java
@@ -9,8 +9,8 @@ import network.crypta.node.Node;
 import network.crypta.node.NodeStats;
 import network.crypta.node.PeerManager;
 import network.crypta.node.PeerNode;
-import network.crypta.node.subsystem.NodeBootstrap;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
+import network.crypta.runtime.bootstrap.NodeBootstrap;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
