+++ src/test/java/network/crypta/runtime/bootstrap/NodeConfigManagerTest.java
@@ -1,4 +1,4 @@
-package network.crypta.node.subsystem;
+package network.crypta.runtime.bootstrap;
 
 import java.nio.file.Path;
 import java.util.Locale;
@@ -10,6 +10,8 @@ import network.crypta.config.Option;
 import network.crypta.config.SubConfig;
 import network.crypta.node.Node;
 import network.crypta.node.ProgramDirectory;
+import network.crypta.node.subsystem.NodeNetworkSubsystem;
+import network.crypta.node.subsystem.NodeStorageSubsystem;
 import network.crypta.runtime.alerts.MeaningfulNodeNameUserAlert;
 import network.crypta.runtime.alerts.UserAlertManager;
 import network.crypta.runtime.services.NodeServicesSubsystem;
