+++ src/test/java/network/crypta/runtime/endpoints/TextModeClientInterfaceServerTest.java
@@ -13,8 +13,8 @@ import network.crypta.crypt.RandomSource;
 import network.crypta.io.NetworkInterface;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
-import network.crypta.node.subsystem.NodeBootstrap;
 import network.crypta.node.subsystem.NodeNetworkSubsystem;
+import network.crypta.runtime.bootstrap.NodeBootstrap;
 import network.crypta.runtime.core.SSL;
 import network.crypta.support.PriorityAwareExecutor;
 import org.jetbrains.annotations.NotNull;
