+++ src/test/java/network/crypta/runtime/endpoints/TextModeClientInterfaceTest.java
@@ -10,7 +10,7 @@ import network.crypta.client.HighLevelSimpleClient;
 import network.crypta.keys.FreenetURI;
 import network.crypta.node.Node;
 import network.crypta.node.NodeClientCore;
-import network.crypta.node.subsystem.NodeBootstrap;
+import network.crypta.runtime.bootstrap.NodeBootstrap;
 import network.crypta.support.Base64;
 import network.crypta.support.io.ArrayBucket;
 import org.junit.jupiter.api.Test;
